Fixed: [SECURITY] (CVE-2024-36104) Path traversal leading to RCE (OFBIZ-13092)
As reported by Leïla, <<if you perform a search on find party screen and try to
sort by partyId, createdDate or else, you qill get this message.
"For security reason this URL is not accepted"
The problem was that the initialURI retrieved by the controlFilter still has
semicolon in it. As the uRIFiltered is removed from its semicolon, the
comparison failed and the error is returned.>>

This was due to the use of URLDecoder::decode below but not on the line
that is now fixed by this commit.
I guess it extends to all such cases; ie URLs that use sorting and such,
possibly using js at some point. I did not get further.