Fixed: [SECURITY] (CVE-2024-36104) Path traversal leading to RCE (OFBIZ-13092)
I've got a strange issue that seems caused by one of this Jira modification:
On demo environment, when clicking on several  sort field links, a
'For security reason this URL is not accepted' error is thrown.

For instance, if you perform a search on find party screen and try to sort by
partyId, createdDate or else, you will get this message.

I tried to reproduce it in a local trunk environment without success : the  
uRIFiltered and the initialURI are the same.

jleroux: but indeed when the same is done on a domain site, for instance
official OFBiz demos, it's reproducible in all demo instances.

This fixes it, with some more: refactoring, formatting and renaming

Thanks: Leïla for report and details.

Conflicts handled by hand (a lot) in ControlFilter.java