Improved: [SECURITY] (CVE-2024-36104) Path traversal leading to RCE (OFBIZ-13092)
Using allowedTokens was a bad idea (when all you have is a (new!) hammer,
everything looks like a nail).
It came to my mind that I already used  StringEscapeUtils::unescapeHtml4
in several places and in this case it's the tool of choice.
Actually I thought about  it when beginning, but was unsure it would be enough.
Now it's clear in my mind and this replaces allowedTokens by simply
StringEscapeUtils::unescapeHtml4

So this also removes the recently added allowedTokens in security.properties.
They are now useless.

Also it's quite better because it handles all cases known or unknown.
Better keep allowedTokens list as short as possible.

Conflicts handled by hand in both files