Commented out repository.apache.org in build.gradle for testing unreleased FreeMarker versions (OFBIZ-12934)
Added the Apache snapshots repository (repository.apache.org) as a commented-out block in build.gradle.
This can be used to test unreleased FreeMarker versions in the future.

Note: The repository is disabled by default, as repository.apache.org is not intended for general-purpose Maven use.
As per https://infra.apache.org/abc, it should only be used for testing pre-production ASF artifacts.