Fixed: Incorrect service call for internal requirement may trigger unnecessary production runs (OFBIZ-13270) (#898)
When an internal requirement is approved, a production run should be
automatically created to fulfill it. Previously, the system invoked a
special service (createProductionRunsForProductBom) instead of the
standard one (createProductionRun). As a result, the system not only
created the production run for the required product, but also additional
runs for all its sub-assemblies, even if those sub-assemblies were
already available in the warehouse or had existing requirements in the
system.