Improved: Set javascriptEnabled to Y by default (OFBIZ-13277)
When a new user is created, by default at the first connexion a user preference javascriptEnabled is set to N if the value weren't initialize before.

With modern framework I suggest to reverse the logic and set it to Y.

A better approach would be removed the preference javascriptEnabled because we can play directly with the theming now, but this represents a lot of work.

So at first step just set Y instead N by default