Improved: Shipping weight doesn't use uom for shipping calculation (OFBIZ-12079)
Currently, you can define on your product a weight with an uom, like 10kg or 200g and you can also define your shipping method estimation cost also with an uom like :
    > to 0kg to 5kg -> 5€
    > to 5kg to 10kg -> 20€
    > to 10kg or more -> 50€

The problem when you calculate you shipping the weight uom isn't use, and you retrieve with a product configure to 200g activate a break 10kg or more

To improve this issue, we need to have access to the dispatcher during the cart so we increase the ShoppingCart signature with dispatcher and keep previous with delegator for backward compatibility. If a shippingCart is created without dispatcher, we generate a new one dedicate to ShoppingCart.