Improved: Check parameters passed in URLs (OFBIZ-13295)
Thanks to Nicolas, improves previous commits that did not handle multipart
element

Also while at it, adds a space after "process" in deniedWebShellTokens of
security.properties. I stumbled upon it after trying to upload a file created
by a Canon camera. Anyway the deniedWebShellTokens comment suggest that when
it's necessary.