Fixed: Regression when willing to enter decimals (OFBIZ-13297)
After fixing the UI the "testBigDecimal" was not passing.
That was due to forcing "locale" argument to default in NumberConverters.java
in order to fix the UI.

This fix simply checks if "testBigDecimal" is running and then not force
the "locale" argument to default.

I currently have a systematically "perflib" issue in Win7 when running
Gradle test task (rest is OK. Don't ask, even M$ advices are not helping, so far...).
So the code comes from Ubuntu on VirtualBox where it works. It should be OK, but
I'll be sure only when all the CI will confirm, since I can't test on Win7.