Fixed: "null DispatchContext" issue under high load and concurrency (OFBIZ-13153)
Under high volumetry system or during test process with server with mutliple core, ofbiz failed some service due to error like :
****
     Caused by: java.lang.NullPointerException: Cannot invoke "org.apache.ofbiz.service.DispatchContext.getSecurity()" because "dctx" is null
****

This is a randomly error who seems to find it origin on the initialisation of the dispatcher. On multiple threading system, the dispatcher object is initialize by twice or more process at the same time that failed the dispatchContext registration.

To solve it, we extract dispatchContext registration during the dispatcher initialisationt and move this function as sync to be ensure that only one thread that set it.

Thanks to Mayank S for raise this issue