Improved: Set uniqueItemName as unique on ScreenWidget (OFBIZ-13315)
When you use a menu item with an iterate-section, the widget engine failed to set a correct unique id

Example :
*****
  <ws:screen name="ListElement">
    <ws:section>
      <ws:iterate-section entry="element" list="elements">
        <ws:section>
          <ws:widgets>
            <ws:include-grid name="ListElementFrom" .../>
            <ws:include-grid name="ListElementTo" .../>
          </ws:widgets>
        </ws:section>
      </ws:iterate-section>
    </ws:section>
  </[code]ws:screen>

  <wf:grid name="ListElementFrom" > ....
     <wf:field name="menuAction" title=" ">
       <wf:include-menu name="ListElementInLineBar"/>
     </wf:field>
  </wf:grid>

  <wf:grid name="ListElementTro" >...
    <wf:field name="menuAction" title=" ">
      <wf:include-menu name="ListElementInLineBar"/>
    </wf:field>
  </wf:grid>

  <wm:menu name="ListElementInLineBar" >
    <wm:menu-item name="delete">
      <wm:link target="Element/delete">...</wm:link>
    </wm:menu-item>
  </wm:menu>
****

When you click on delete link, you have twice the submission, each menu present because the menu id is the same.

The reason come from the process to build a unique name use on the js script to submit.

MacroMenuRenderer.java:223 where we concat the menu name, menu item name and different index present in context ('menuUniqueItemIndex', 'itemIndex' and 'parentItemIndex'). It's possible if we manage different index on this screen to escape this issue, but after ready and taking a step back, this will be increase the screen complexity for nothing link to this problematic.

We simplify it to generate each time a real unique name that we can prefix and if we need to spot an element from a custom js script I suggest to use index instead name and a unique name to build correctly through multiple context index.

Follow this way, we introduce a new class UtilRandom to centralize methods to manage a unique id/name and propagate this logic on all unique item name using on the screen widget engine.