Fixed: Compound's single form are loaded as ModelGrid in memory (OFBIZ-13330)
If you create a simple single form in a compound, when OFBiz load it in memory, the element create isn't a ModelForm but a ModelGrid.

This came from the object builder that analyze the XML tag name with the name space. In FormFactory::createModelForm

*******

if ("form".equals(formElement.getTagName()) && ...)
{     return new ModelSingleForm( ...) }

return new ModelGrid( ...)

*******

Or :  "wf:form" != "form"

Simple fix to remove the namespace from the analysis.