Improved: Manage marital status type history for person (OFBIZ-13329)
on the framework we can add a marital status to a person through enumeration on entity Person.

But we can't manage the history of change. After scan the data model book, we proposed to improve party model entity with two new entities :

    MaritalStatus : to stock the history of marital status change
    MaritalStatusType : like Len Silverton suggest on his book to isolate these types.

We move the current field maritalStatusEnumId to deprecated and introduce a migration service to move field maritalStatusEnumId to maritalStatusTypeId

Thanks to Charles Steltzlen for this improvement