Fixed: Form under ModelGrid don't use values in parameters and block others (OFBiz-13331)
When you load a widget screen, to know if a form collect some values on parameters map, a boolean useRequestParameters is present on context.

If the model form detect that it build a ModelGrid it set useRequestParameters at false.

Problem if some form need to collect information after a grid, we need to set it to true.

To solve this we add the form name in context that not collect on parameters to don't impact some others.

By the way, we reformat a few old code with new java language logical